Inline FeedbackHandler in FeedbackWorker
